refactor: 계층형 아키텍처에 맞는 패키지명으로 수정
