+++ src/main/java/roomescape/business/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.business;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
