+++ src/main/java/roomescape/business/dto/ReservationTimeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.domain.dto;
+package roomescape.business.dto;
 
 import java.time.LocalTime;
