+++ src/main/java/roomescape/persistence/GeneralRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repositiory;
+package roomescape.persistence;
 
 import java.util.List;
