+++ src/main/java/roomescape/presentation/ReservationQueryController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.presentation;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.ResponseBody;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.service.ReservationService;
+import roomescape.business.Reservation;
+import roomescape.business.ReservationTime;
+import roomescape.business.service.ReservationService;
 
 @Controller
 public class ReservationQueryController {
