+++ src/test/java/roomescape/MissionStepTest.java
@@ -17,9 +17,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationCommandController;
-import roomescape.controller.ReservationQueryController;
-import roomescape.domain.Reservation;
+import roomescape.presentation.ReservationCommandController;
+import roomescape.presentation.ReservationQueryController;
+import roomescape.business.Reservation;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
