+++ src/test/java/roomescape/persistence/ReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.repositiory;
+package roomescape.persistence;
 
 import java.time.LocalTime;
 import java.util.List;
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.domain.ReservationTime;
+import roomescape.business.ReservationTime;
 
 @JdbcTest
 @Import(ReservationTimeRepository.class)
